--- vue3-ts-quasar/src/components/RepoCard/index.ts
@@ -1,2 +1,0 @@
-import RepoCard from './RepoCard.vue';
-export default RepoCard;
