--- vue3-ts-quasar/src/components/UserProfileCard/index.ts
@@ -1,2 +1,0 @@
-import UserProfileCard from './UserProfileCard.vue';
-export default UserProfileCard;
