+++ vue3-ts-quasar/tests/unit/components/NavHeader.spec.ts
@@ -1,5 +1,5 @@
 import { shallowMount } from '@vue/test-utils';
-import NavHeader from '@/components/NavHeader/NavHeader.vue';
+import { NavHeader } from '@/components';
 import { createTestingPinia } from '@pinia/testing';
 
 describe('Logo', () => {
